Support specifying the database to use for the store_result method
This change allows specifying the database to using when storing a
result to the database.

I use this change to update a celery task while it is working
on Django models in a transaction. With this change, I can use a
separate connection to update the task without waiting for the
transaction to commit.

This change is related to https://github.com/celery/celery/pull/5017